refcto almost done but something wrong to check
