+++ srcs/parsing/parsing_utils.c
@@ -6,58 +6,11 @@
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/25 16:53:40 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 15:13:07 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 15:44:05 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	create_new_cmd_list_node(t_cmd **node)
-{
-	t_cmd	*new_elem;
-	t_cmd	*ptr;
-
-	if (!(*node) || !(node))
-		return ;
-	ptr = *node;
-	while (ptr->next)
-		ptr = ptr->next;
-	new_elem = ft_calloc(sizeof(t_cmd), 1);
-	if (!new_elem)
-		return ;
-	new_elem->input = STDIN_FILENO;
-	new_elem->output = STDOUT_FILENO;
-	ptr->next = new_elem;
-	*node = (*node)->next;
-}
-
-int	count_token_cmd(t_minish *data, int *i)
-{
-	int	count;
-	int	temp;
-
-	temp = *i;
-	count = 0;
-
-	while (data->tokens[temp] && data->tokens[temp][0] != '|'
-		&& data->tokens[temp][0] != '\0' && data->tokens[temp][0] != '<'
-		&& data->tokens[temp][0] != '>')
-	{
-		count++;
-		temp++;
-	}
-	return (count);
-}
-
-int	is_builtin(char *cmd)
-{
-	if (!ft_strncmp(cmd, "unset", 5) || !ft_strncmp(cmd, "env", 3)
-		|| !ft_strncmp(cmd, "pwd", 3) || !ft_strncmp(cmd, "export", 6)
-		|| !ft_strncmp(cmd, "cd", 2) || !ft_strncmp(cmd, "echo", 4)
-		|| !ft_strncmp(cmd, "exit", 4))
-		return (1);
-	return (0);
-}
 
 int	heredoc_handling(t_minish *data, t_cmd **node, int *i)
 {
@@ -65,7 +18,7 @@ int	heredoc_handling(t_minish *data, t_cmd **node, int *i)
 
 	if (!data->tokens[*i + 1])
 	{
-		// error_manager(11, data, NULL);
+		printf("Error manager to be set\n");
 		return (1);
 	}
 	else
@@ -89,7 +42,7 @@ int	input_redirection(t_minish *data, t_cmd **node, int *i)
 {
 	if (!data->tokens[*i + 1])
 	{
-		// error_manager(8, data, NULL);
+		printf("Error manager to be set\n");
 		return (1);
 	}
 	else
@@ -101,7 +54,6 @@ int	input_redirection(t_minish *data, t_cmd **node, int *i)
 			*i += 1;
 			return (1);
 		}
-			// error_manager(9, data, NULL);
 		*i += 1;
 	}
 	return (0);
@@ -111,7 +63,7 @@ int	output_redirection(t_minish *data, t_cmd **node, int *i)
 {
 	if (!data->tokens[*i + 1])
 	{
-		// error_manager(8, data, NULL);
+		printf("Error manager to be set\n");
 		return (1);
 	}
 	else
@@ -133,7 +85,7 @@ int	output_append_redirection(t_minish *data, t_cmd **node, int *i)
 {
 	if (!data->tokens[*i + 1])
 	{
-		// error_manager(8, data, NULL);
+		printf("Error manager to be set\n");
 		return (1);
 	}
 	else
@@ -167,34 +119,5 @@ int	pipe_new_node(t_minish *data, t_cmd **node, int *i)
 		*i += 1;
 		return (0);
 	}
-}
-
-int	cmds_number(t_minish *data)
-{
-	int		count;
-	t_cmd	*ptr;
-
-	count = 0;
-	ptr = data->cmds;
-	while (ptr)
-	{
-		count++;
-		ptr = ptr->next;
-	}
-	return (count);
-}
-
-int	check_pipes(t_minish *data)
-{
-	int	i;
-
-	i = 0;
-	while (data->tokens[i])
-	{
-		if (data->tokens[i][0] == '|')
-			return (1);
-		i++;
-	}
-	return (0);
 }
