+++ srcs/settings_utils/memory_stuff_2.c
@@ -6,7 +6,7 @@
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/06 12:03:25 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 15:13:47 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 15:47:17 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,6 @@ void	free_linked_list_full_cmd(t_minish *data)
 		ptr = ptr->next;
 		free(tmp);
 	}
-	// free(ptr);
 	data->cmds = NULL;
 }
 
@@ -103,22 +102,4 @@ void	free_tokens(t_minish *data)
 	}
 }
 
-void	exit_clean(t_minish *data)
-{
-	if (data)
-	{
-		if (data->input)
-			free(data->input);
-		if (data->envp)
-			dict_free(&data->envp);
-		if (data->tokens)
-			free_tokens(data);
-		if (data->cmds)
-			free_linked_list_full_cmd(data);
-		free_path_dir(data);
-		free_env_table(data);
-		if (data->aux)
-			ft_lstclear(&data->aux, free);
-		free(data);
-	}
-}
+
