+++ srcs/settings_utils/settings_2.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   settings.c                                         :+:      :+:    :+:   */
+/*   settings_2.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/06 17:43:22 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 15:13:52 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 15:51:11 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,7 +88,6 @@ char	**tab_envp_updated(t_minish *data)
 	{
 		join = ft_strjoin(ptr->key, "=");
 		data->env_table[n_var] = ft_strjoin(join, ptr->value);
-		// printf("data->env_table[n_var] = #%s#\n", data->env_table[n_var]);
 		free(join);
 		n_var++;
 		ptr = ptr->next;
@@ -114,7 +113,7 @@ void	init_cmd(t_minish *data)
 	data->cmds = command;
 }
 
-static void	init_ptrs(t_minish *data, char *envp[])
+void	init_ptrs(t_minish *data, char *envp[])
 {
 	data->path = NULL;
 	data->path_dir = NULL;
@@ -124,37 +123,4 @@ static void	init_ptrs(t_minish *data, char *envp[])
 	data->env_table = NULL;
 	data->envp = dup_envp(envp);
 	data->env_table = NULL;
-}
-
-t_minish	*init_data(char *envp[])
-{
-	t_minish	*data;
-
-	data = malloc(sizeof(*data));
-	if (!data)
-	{
-		ft_printf("ERROR ALLOCATING DATA: %s\n", strerror(errno));
-		exit(1);
-	}
-	data->n_cmd = 0;
-	data->n_tokens = 0;
-	data->btw_double_quotes = 0;
-	data->btw_simple_quotes = 0;
-	data->start = 0;
-	data->end = 0;
-	init_cmd(data);
-	init_ptrs(data, envp);
-	return (data);
-}
-
-void	update_envp(t_dict *envp)
-{
-	int		lvl;
-	t_dict	*ptr;
-
-	set_varvalue(envp, "SHELL", 5, "minishell");
-	ptr = dict_findvar(envp, "SHLVL", 5);
-	lvl = ft_atoi(ptr->value) + 1;
-	free(ptr->value);
-	ptr->value = ft_itoa(lvl);
 }
