+++ srcs/user/prompt.c
@@ -6,7 +6,7 @@
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 11:07:01 by mpagani           #+#    #+#             */
-/*   Updated: 2023/04/01 15:14:10 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 16:33:32 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,6 +40,7 @@ void	setting_prompt(t_minish *data)
 			tab_envp_updated(data);
 			if (!lexer_input(data))
 				executing_commands(data);
+			printf("COUCOU\n");
 		}
 		free(prefix);
 		if (data->input == NULL)
