+++ srcs/user/signal.c
@@ -6,13 +6,11 @@
 /*   By: mpagani <mpagani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/06 19:51:32 by fbelfort          #+#    #+#             */
-/*   Updated: 2023/04/01 15:14:40 by mpagani          ###   ########.fr       */
+/*   Updated: 2023/04/01 16:30:25 by mpagani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	handle_ctrlc_exec(int sign);
 
 void	handle_ctrlc(int sign, siginfo_t *info, void *context)
 {
@@ -42,25 +40,21 @@ void	handle_ctrld_exec(int sign)
 {
 	printf("Quit (core dumped)\n");
 	g_status = 128 + sign;
-	// exit(g_status);
 }
 
 void	handle_ctrlc_exec(int sign)
 {
 	write(1, "\n", 1);
 	g_status = 128 + sign;
-	// signal(SIGINT, SIG_IGN);
-	// exit(g_status);
 }
 
 void	handle_ctrlc_heredoc(int sign)
 {
 	write(1, "\n", 1);
 	g_status = 128 + sign;
-	// printf("g_status = %d\n", g_status);
 	close(STDIN_FILENO);
-	// exit(g_status);
 }
+
 
 /**
  * @brief
@@ -84,6 +78,7 @@ void	set_signals(int caller)
 		sigaction(SIGINT, &sa, NULL);
 		sigaction(SIGQUIT, &sa, NULL);
 	}
+		// signal_caller_prompt(sa);
 	if (caller == EXEC)
 	{
 		signal(SIGINT, SIG_IGN);
@@ -92,6 +87,7 @@ void	set_signals(int caller)
 		sa.sa_handler = handle_ctrld_exec;
 		sigaction(SIGQUIT, &sa, NULL);
 	}
+		// signal_caller_exec(sa);
 	if (caller == HEREDOC)
 	{
 		sa.sa_handler = handle_ctrlc_heredoc;
@@ -100,3 +96,11 @@ void	set_signals(int caller)
 	if (caller == OFF)
 		signal(SIGINT, SIG_IGN);
 }
+
+// void	signal_caller_prompt(struct sigaction sa)
+// {
+// }
+
+// void	signal_caller_exec(struct sigaction sa)
+// {
+// }
